pyproject.toml
[project]
name = "infai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "azure-search-documents>=11.6.0",
    "dotenv>=0.9.9",
    "pydantic-ai-slim[openai]>=1.46.0",
    "pytest-asyncio>=0.23.3",
    "qdrant-client>=1.16.1",
    "rich>=14.2.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "sqlmodel>=0.0.27",
    # Delete this when not using a local stdio MCP server directly from this repo
    "fastmcp>=2.14.4",
]

[dependency-groups]
dev = [
    "pre-commit>=4.4.0",
    "pyright>=1.1.407",
    "pytest>=9.0.0",
    "pytest-asyncio>=0.23.3",
    "requests>=2.32.5",
    "ruff>=0.14.4",
    # "ty>=0.0.2", # Switch to ty when they support reportMatchNotExhaustive
]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
